requirements.txt
fastapi==0.95.0
uvicorn==0.18.2
snowflake-connector-python==3.0.0
python-dotenv==0.20.0